Refactor server code structure
